refactor: SendNotificationCommand 패키지 변경
